feat(app): add legacy top page, chanari routes, and route splitting (#49)
- お気楽チャット旧トップページを追加し、各部屋リンクを内部ルート化
- 左カラム全ルームを room 定義に追加し、全ルームを enabled 化
- Supabase の直近発言から部屋別参加人数を集計して表示
- 画像依存のヘッダーをテキスト/CSSベースのヘッダーへ置き換え
- /chanari/:roomId のレトロUIとルーティングを追加
- chanari リンク、管理者部屋、問い合わせ/お知らせリンクを反映
- App のルート描画を Chat / Top / Chanari / NotFound route に分割
- 初期表示時の全体ローディング fallback を外し、各ルートを即時表示
- 重複していた preloadCriticalResources を削除
- BASE_URL を使うようトップタブ遷移を修正
- ルーティング、トップページ、人数集計、App のテストを更新
