+++ src/routes/ChanariRoute.tsx
@@ -1,0 +1,3 @@
+import ChanariChatPage from '@features/chanari-chat/ChanariChatPage';
+
+export default ChanariChatPage;
