+++ src/routes/NotFoundRoute.tsx
@@ -1,0 +1,3 @@
+import NotFoundPage from '../pages/NotFoundPage';
+
+export default NotFoundPage;
