+++ src/routes/TopRoute.tsx
@@ -1,0 +1,3 @@
+import TopPage from '@features/top/TopPage';
+
+export default TopPage;
